windows/requirements.txt
# Dependencias para Windows Network Optimizer

# GUI moderna con CustomTkinter (tema oscuro nativo)
customtkinter>=5.2.0
Pillow>=10.0.0

# Gráficas en tiempo real
matplotlib>=3.8.0

# Monitoreo de sistema
psutil>=5.9.0

# Base de datos ligera (JSON puro, sin SQL)
tinydb>=4.8.0

# Notificaciones
winotify==1.1.0

# Logging mejorado
colorlog==6.8.2

# Para compilar a .exe
pyinstaller>=6.0.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
